[feat]: improvements to avatar
