+++ src/components/CategoriesListItem/CategoriesListItem.tsx
@@ -1,7 +1,4 @@
-import React, { FC, useEffect, useState } from "react";
-import firebase from "firebase";
-
-import Avatar from "@material-ui/core/Avatar";
+import React, { FC } from "react";
 import IconButton from "@material-ui/core/IconButton";
 import ListItem from "@material-ui/core/ListItem";
 import ListItemAvatar from "@material-ui/core/ListItemAvatar";
@@ -11,7 +8,7 @@ import ListItemSecondaryAction from "@material-ui/core/ListItemSecondaryAction";
 import DeleteIcon from "@material-ui/icons/DeleteOutline";
 
 import CategoriesListItemProps from "./types";
-import { useNotificationDispatch } from "../../contexts/NotificationProvider";
+import CategoryIconAvatar from "../CategoryIconAvatar";
 
 const CategoriesListItem: FC<CategoriesListItemProps> = ({
   onRemove,
@@ -19,53 +16,14 @@ const CategoriesListItem: FC<CategoriesListItemProps> = ({
   onAvatarClick = () => {},
   type
 }) => {
-  const [avatarIcon, setAvatarIcon] = useState("");
-  const dispatchNotifications = useNotificationDispatch();
-
-  useEffect(() => {
-    async function getIcon() {
-      if (category.icon) {
-        const firebaseRef = firebase
-          .storage()
-          .refFromURL(`gs://muhnee-app.appspot.com/${category.icon}`);
-        try {
-          setAvatarIcon(await firebaseRef.getDownloadURL());
-        } catch (e) {
-          console.error(e);
-          setAvatarIcon("");
-          if (e.code === "storage/object-not-found") {
-            dispatchNotifications({
-              type: "@@NOTIFICATION/PUSH",
-              notification: {
-                message: `Cannot find receipt`,
-                type: "error"
-              }
-            });
-          } else {
-            dispatchNotifications({
-              type: "@@NOTIFICATION/PUSH",
-              notification: {
-                message: `an unknown error occurred.`,
-                type: "error"
-              }
-            });
-          }
-        }
-      } else {
-        setAvatarIcon("");
-      }
-    }
-    getIcon();
-  }, [avatarIcon, category.icon, dispatchNotifications, setAvatarIcon]);
-
   return (
     <ListItem key={category.id}>
       <ListItemAvatar>
-        <Avatar
-          src={avatarIcon}
+        <CategoryIconAvatar
           onClick={() => {
             onAvatarClick(type, category.id, category.name);
           }}
+          category={category}
         />
       </ListItemAvatar>
       <ListItemText primary={category.name} />
