+++ src/components/CategoryIconAvatar/index.ts
@@ -1,0 +1,2 @@
+import CategoryIconAvatar from "./CategoryIconAvatar";
+export default CategoryIconAvatar;
